spacetimedb_query_builder::expr

Trait RHS

pub trait RHS<T, V> {
    // Required method
    fn to_expr(self) -> Operand<T>;
}

Trait for types that can be used as the right-hand side of a comparison with a column of type V in table T.

This trait is implemented for Col<T, V> and various literal types.

Required Methods

fn to_expr(self) -> Operand<T>